requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.14.0
pytz==2021.3
sqlparse==0.4.2
Unipath==1.1
whitenoise==5.3.0
pandas==1.4.1
numpy==1.22.2
openpyxl==3.0.9
psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1,<2.1